Major rework to handling of font metrics.
Add an autoscale attribute to various classes; this controls whether
metrics are to be passed and returned as scaled floats, being
hb_position_t integer values divided by 64 (True), or as the original
unscaled integers (False). This attribute is automatically set to True
for Faces created from FreeType faces, but caller now has to specify
this when creating other kinds of Face objects. Also it has to be
specified for FontFuncs objects, to control how to interpret metrics
returned from callback routines.

Buffer objects also have this attribute, to remember the setting from
the Font object passed to the last shape operation on the buffer. This
way they can return the correctly-scaled glyph metrics.